--- lib/puppet/provider/reboot/base.rb
@@ -1,18 +1,0 @@
-require 'puppet/type'
-
-Puppet::Type.type(:reboot).provide(:base) do
-  def self.instances
-    []
-  end
-
-  def when
-    :absent
-  end
-
-  def when=(value)
-  end
-
-  def cancel_transaction
-    Puppet::Application.stop!
-  end
-end
